Moved test files and changed jest config to ignore server tests for circleCI
