+++ package.json
@@ -63,6 +63,6 @@
       "json",
       "jsx"
     ],
-    "testEnvironment": "node"
+    "coveragePathIgnorePatterns": ["<rootDir>/server/"]
   }
 }
